fix(ai-v2): fix build breaks from PR #85 merge (CS1503/CS0103/CS0136)
- AviationCombatEstimator: cast UnitData.Attack (int) to float before the
  IReadOnlyList<float> overload, otherwise overload resolution fell through
  to the IReadOnlyList<UnitData> overload and failed to convert.
- WorldAnalysis.Economy: drop dangling TaskScoreDiagnostics.Log call — that
  type doesn't exist anywhere in the codebase, tracing-only, not load-bearing.
- AggressionMissionPlanner.ReinforcementCandidate: rename the unpinned-branch
  `projection` local to `unpinnedProjection` — it collided with the same name
  declared later in the enclosing method scope (CS0136).

Verified with `dotnet build Assembly-CSharp.csproj` — 0 errors, 0 warnings.

+++ Assets/Scripts/Combat/AviationCombatEstimator.cs
@@ -93,7 +93,7 @@ namespace Game.Combat
             float knownDefense, float knownAttack,
             IReadOnlyList<WorthIt.DefenderProfile> knownDefenders, AirStrikePolicy policy) =>
             EstimateAirStrike(
-                aircraft?.Where(x => x != null).Select(x => x.Attack).ToList(),
+                aircraft?.Where(x => x != null).Select(x => (float)x.Attack).ToList(),
                 knownDefense, knownAttack, knownDefenders, policy);
 
         // Snapshot-safe overload for strategic planning. The live UnitData overload above and
